requirements.txt
aenum==3.1.15
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==5.1.1
diskcache==5.6.3
evadb==0.3.9
filelock==3.13.1
fsspec==2023.10.0
greenlet==3.0.1
huggingface-hub==0.19.4
idna==3.5
Jinja2==3.1.2
lark==1.1.8
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
psutil==5.9.6
psycopg2==2.9.9
py==1.11.0
pydantic==1.10.13
PyPDF2==3.0.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
retry==0.9.2
safetensors==0.4.0
six==1.16.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
sympy==1.12
thefuzz==0.20.0
tokenizers==0.15.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
